Add files via upload
Subjects in a GSS were asked their opinions about government spending on the environment, health, assistance to big cities, and law enforcement. Each participant evaluated the government spending on the four categories by answering whether the government is spending not enough, about right, or too much. The data in this study collected opinions from 597 participants. From this data, we performed a homogeneous association model to analyze the association between the opinions on environment spending and health spending. We then  looked to see what other relevant associations we can find in the data.